Attempt at simplifying the code.
MemQueue do not hold a reference to the last file that
records their position anymore.

Instead, we just rely on the fact that we always log the position
of empty mem queues before deleting a file.

Also contrary to the previous behavior, we do not do that
always right before GC. We do this only if GC will actually
trigger the deletion of at least one file.