Moving toward a paging-based recordlog
Instead of chained list of 128mb files, we
pre-allocate a file, and split it into pages.

A header in the file stores the order in which the pages
should be read.

On the write side, when we truncate queues, we often end up
freeing pages. When all pages are in use, we perform a GC.
It simply reorders pages, and updates the file header accordingly.